Move reference code, add url support
